+++ Reinvest/Notifications/src/Port/Api/StoredEventsController.ts
@@ -2,18 +2,22 @@ import { Pagination, UUID } from 'HKEKTypes/Generics';
 import { AccountActivitiesRepository } from 'Notifications/Adapter/Database/Repository/AccountActivitiesRepository';
 import { StoredEventRepository } from 'Notifications/Adapter/Database/Repository/StoredEventRepository';
 import { ProcessStoredEvent } from 'Notifications/Application/UseCase/ProcessStoredEvent';
+import { PushNotificationRepository } from 'Notifications/Adapter/Database/Repository/PushNotificationRepository';
 
 export class StoredEventsController {
   private accountActivitiesRepository: AccountActivitiesRepository;
+  private pushNotificationRepository: PushNotificationRepository;
 
   constructor(
     storedEventsRepository: StoredEventRepository,
     processStoredEventUseCase: ProcessStoredEvent,
     accountActivitiesRepository: AccountActivitiesRepository,
+    pushNotificationRepository: PushNotificationRepository,
   ) {
     this.storedEventsRepository = storedEventsRepository;
     this.processStoredEventUseCase = processStoredEventUseCase;
     this.accountActivitiesRepository = accountActivitiesRepository;
+    this.pushNotificationRepository = pushNotificationRepository;
   }
 
   static getClassName = () => 'StoredEventsController';
@@ -32,5 +36,9 @@ export class StoredEventsController {
     const activities = await this.accountActivitiesRepository.listActivities(profileId, accountId, pagination);
 
     return activities.map(activity => activity.getView());
+  }
+
+  async registerPushNotificationDevice(profileId: UUID, deviceId: UUID): Promise<boolean> {
+    return this.pushNotificationRepository.registerDevice(profileId, deviceId);
   }
 }