+++ serverless-functions.ts
@@ -14,6 +14,7 @@ import { CronDocumentSyncFunction, CronDocumentSyncResources } from './devops/fu
 import { CronNotificationsFunction, CronNotificationsResources } from './devops/functions/cron/notifications/cron-notifications-config';
 import { CronVendorsSyncFunction, CronVendorsSyncResources } from './devops/functions/cron/vendorsSync/cron-vendors-sync-config';
 import { ExplorerLambdaFunction, ExplorerLambdaResources } from './devops/functions/explorer/explorer-config';
+import { FirebaseFunction, FirebaseResources } from './devops/functions/firebase/queue-config';
 import { MigrationLambdaFunction, MigrationLambdaResources } from './devops/functions/migration/migration-config';
 import { PdfGeneratorFunction, PdfGeneratorResources } from './devops/functions/pdfGenerator/queue-config';
 import { cognitoPostSignUpFunction, CognitoPostSignUpResources } from './devops/functions/postSignUp/postSignUp-config';
@@ -47,6 +48,7 @@ const serverlessConfiguration: AWS = {
       LocalCognitoClientId: { Ref: 'LocalCognito' },
       SQS_QUEUE_URL: getAttribute('SQSNotification', 'QueueUrl'),
       SQS_PDF_GENERATOR_URL: getAttribute('SQSPdfGenerator', 'QueueUrl'),
+      SQS_FIREBASE_QUEUE_URL: getAttribute('SQSFirebase', 'QueueUrl'),
       EMAIL_SEND_FROM: '${env:EMAIL_SEND_FROM}',
       EMAIL_REPLY_TO: '${env:EMAIL_REPLY_TO}',
       WEB_APP_URL: '${env:WEB_APP_URL}',
@@ -67,6 +69,7 @@ const serverlessConfiguration: AWS = {
       DEALPATH_API_URL: '${env:DEALPATH_API_URL}',
       DEALPATH_AUTHORIZATION_TOKEN: '${env:DEALPATH_AUTHORIZATION_TOKEN}',
       DEALPATH_VERSION_HEADER: '${env:DEALPATH_VERSION_HEADER}',
+      FIREBASE_SERVICE_ACCOUNT_JSON: '${env:FIREBASE_SERVICE_ACCOUNT_JSON}',
     },
     apiGateway: {
       minimumCompressionSize: 1024,
@@ -99,6 +102,7 @@ const serverlessConfiguration: AWS = {
     cognitoPreSignUpFunction,
     tests: TestsFunction,
     pdfGenerator: PdfGeneratorFunction,
+    firebase: FirebaseFunction,
   },
   resources: {
     Description: 'REINVEST ${sls:stage} API functions',
@@ -118,6 +122,7 @@ const serverlessConfiguration: AWS = {
       ...CronDividendsCalculationResources,
       ...CronDividendsDistributionResources,
       ...PdfGeneratorResources,
+      ...FirebaseResources,
       ...CronNotificationsResources,
     },
     Outputs: {