+++ serverless-local.ts
@@ -14,6 +14,7 @@ import { CronDocumentSyncFunction, CronDocumentSyncResources } from './devops/fu
 import { CronNotificationsFunction, CronNotificationsResources } from './devops/functions/cron/notifications/cron-notifications-config';
 import { CronVendorsSyncFunction, CronVendorsSyncResources } from './devops/functions/cron/vendorsSync/cron-vendors-sync-config';
 import { ExplorerLambdaFunction, ExplorerLambdaResources } from './devops/functions/explorer/explorer-config';
+import { FirebaseFunction, FirebaseResources } from './devops/functions/firebase/queue-config';
 import { MigrationLambdaFunction, MigrationLambdaResources } from './devops/functions/migration/migration-config';
 import { PdfGeneratorFunction, PdfGeneratorResources } from './devops/functions/pdfGenerator/queue-config';
 import { cognitoPostSignUpFunction, CognitoPostSignUpResources } from './devops/functions/postSignUp/postSignUp-config';
@@ -47,6 +48,7 @@ const serverlessConfiguration: AWS = {
       ApiUrl: 'http://localhost:3000/api',
       SQS_QUEUE_URL: 'http://localhost:9324/000000000000/development-sqs-notification',
       SQS_PDF_GENERATOR_URL: 'http://localhost:9324/000000000000/development-sqs-pdf-generator',
+      SQS_FIREBASE_QUEUE_URL: 'http://localhost:9324/000000000000/development-sqs-firebase',
       IT_IS_LOCAL: 'true',
     },
     logs: {
@@ -80,6 +82,7 @@ const serverlessConfiguration: AWS = {
     unauthorizedEndpoints: UnauthorizedEndpointsFunction,
     tests: TestsFunction,
     pdfGenerator: PdfGeneratorFunction,
+    firebase: FirebaseFunction,
   },
   resources: {
     Resources: {
@@ -102,6 +105,7 @@ const serverlessConfiguration: AWS = {
       ...CronDividendsCalculationResources,
       ...CronDividendsDistributionResources,
       ...PdfGeneratorResources,
+      ...FirebaseResources,
       ...CronNotificationsResources,
     },
     Outputs: {
@@ -136,6 +140,7 @@ const serverlessConfiguration: AWS = {
     'serverless-offline': {
       useChildProcesses: true,
       noPrependStageInUrl: true,
+      disableScheduledEvents: true,
     },
     // 'serverless-offline-watcher': [
     //   {